+++ changelog.md
@@ -20,7 +20,7 @@ slide:
     column:
       nbText: "Right column"
 ```
-- `footer` template to a footer to every slide. The input should be HTML.
+- `footer` template to a footer to every slide. The input is a string with markdown, unless the argument `rawHtml=true` in which case it is treated as raw HTML.
   - The footer will be shown on all slides except those with image-, video- or iframe-backgrounds.
 - Experimental features - the API for these are not yet stable and might change in the future:
   - `fragmentThen` - Allows the construction of fragments like `growThenShrink` by `fragmentThen(grows, shrinks)`. The second animations happens at the same time as the next animation, reducing the number of clicks needed.